requirements.txt
django-filebrowser<4
django-filter
django-grappelli<3
django-redis
django-rq
django>=3.2,<4
djangorestframework
drf-extensions
pillow
psycopg2-binary
django-environ==0.11.2
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3